requirements/local.txt
# Local development dependencies go here
-r base.txt
django-debug-toolbar==4.2.0
django-extensions==3.2.3
collection-json==0.1.1
coverage==7.3.1
pylint==2.17.5
flake8==6.1.0
isort==5.12.0
pudb==2022.1.3
https://github.com/msbrogli/rpudb/archive/master.zip
